requirements.txt
pygame
keras
tensorflow
pandas
seaborn
numpy
neat-python
